+++ src/interface/aimrt_module_cpp_interface/executor/timer.h
@@ -4,41 +4,12 @@
 #pragma once
 
 #include <chrono>
-#include <type_traits>
 
 #include "aimrt_module_cpp_interface/executor/executor.h"
 #include "exception.h"
+#include "same_arg_trait.h"
 
 namespace aimrt::executor {
-
-template <typename T>
-struct function_args;
-
-template <typename R, typename... Args>
-struct function_args<R(Args...)> {
-  using type = std::tuple<Args...>;
-};
-
-template <typename R, typename C, typename... Args>
-struct function_args<R (C::*)(Args...)> {
-  using type = std::tuple<Args...>;
-};
-
-template <typename R, typename C, typename... Args>
-struct function_args<R (C::*)(Args...) const> {
-  using type = std::tuple<Args...>;
-};
-
-template <typename F>
-struct function_args {
-  using type = typename function_args<decltype(&F::operator())>::type;
-};
-
-template <typename F>
-using function_args_t = typename function_args<F>::type;
-
-template <typename F1, typename F2>
-concept SameArguments = std::same_as<function_args_t<F1>, function_args_t<F2>>;
 
 class TimerBase {
  public:
@@ -86,8 +57,8 @@ class TimerBase {
 };
 
 template <typename TaskType>
-  requires(SameArguments<TaskType, std::function<void()>> ||
-           SameArguments<TaskType, std::function<void(TimerBase&)>>)
+  requires(common::util::SameArguments<TaskType, std::function<void()>> ||
+           common::util::SameArguments<TaskType, std::function<void(TimerBase&)>>)
 class Timer : public TimerBase {
  public:
   Timer(ExecutorRef executor, std::chrono::nanoseconds period, TaskType&& task, bool auto_start = true)
